move some things to better places
