+++ src/main/kotlin/net/blockventuremc/modules/rides/track/segments/AccelerationSegment.kt
@@ -1,7 +1,6 @@
 package net.blockventuremc.modules.rides.track.segments
 
-import net.blockventuremc.modules.rides.track.TrackNode
-import net.blockventuremc.modules.structures.Train
+import net.blockventuremc.modules.structures.impl.Train
 
 /**
  * Represents a segment of acceleration on a track.
