+++ src/main/kotlin/net/blockventuremc/modules/rides/track/segments/LiftSegment.kt
@@ -1,8 +1,6 @@
 package net.blockventuremc.modules.rides.track.segments
 
-import net.blockventuremc.modules.rides.track.TrackNode
-import net.blockventuremc.modules.rides.track.utils.calculateAdjustedSpeed
-import net.blockventuremc.modules.structures.Train
+import net.blockventuremc.modules.structures.impl.Train
 import kotlin.math.abs
 
 /**
