+++ src/main/kotlin/net/blockventuremc/modules/rides/track/segments/StationSegment.kt
@@ -1,6 +1,6 @@
 package net.blockventuremc.modules.rides.track.segments
 
-import net.blockventuremc.modules.structures.Train
+import net.blockventuremc.modules.structures.impl.Train
 
 open class StationSegment(startId: Int,endId: Int, val stationSpeed: Float) : TrackSegment(startId, endId) {
