+++ src/main/kotlin/net/blockventuremc/modules/structures/StructureManager.kt
@@ -1,6 +1,7 @@
 package net.blockventuremc.modules.structures
 
 import net.blockventuremc.VentureLibs
+import net.blockventuremc.modules.`fun`.baloon.Balloon
 import org.bukkit.entity.Player
 import java.io.Closeable
 import java.util.UUID
