+++ src/main/kotlin/net/blockventuremc/modules/structures/commands/CustomEntityCommand.kt
@@ -1,6 +1,8 @@
-package net.blockventuremc.modules.structures
+package net.blockventuremc.modules.structures.commands
 
 import net.blockventuremc.annotations.VentureCommand
+import net.blockventuremc.modules.structures.*
+import net.blockventuremc.modules.structures.impl.Seat
 import net.blockventuremc.utils.itembuilder.ItemBuilder
 import org.bukkit.Material
 import org.bukkit.command.Command
@@ -10,7 +12,6 @@ import org.bukkit.entity.Player
 import org.bukkit.inventory.ItemStack
 import org.bukkit.permissions.PermissionDefault
 import org.bukkit.util.Vector
-import java.util.UUID
 
 @VentureCommand(
     name = "customentity",
